Fix ERROR Platform.channels: OpenCV(4.12.0) D:\a\opencv-python\opencv-python\opencv\modules\imgproc\src\resize.cpp:4086: error: (-215:Assertion failed) func != 0 in function 'cv::hal::resize'

+++ modules/Models/Segmentation/Segmentor.py
@@ -497,7 +497,7 @@ class SegformerSegmentor(BaseSegmentor):
         confidence_map = None
         if self.return_confidence:
             softmax_probs = torch.softmax(upsampled_logits, dim=1)
-            confidence_map = torch.max(softmax_probs, dim=1)[0].cpu().numpy()[0]
+            confidence_map = torch.max(softmax_probs, dim=1)[0].to(torch.float32).cpu().numpy()[0]
 
         # Extract individual masks and bounding boxes for each detected class
         instances = []
